% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_users.R
\name{create_test_user}
\alias{create_test_user}
\title{Create a test user.}
\usage{
create_test_user(site_id, authorization = Sys.getenv("MELI_ACCESS_TOKEN"))
}
\arguments{
\item{site_id}{Required. The country in which you want to operate.}

\item{authorization}{Required. A valid access token from developers.mercadolibre.com. See the \href{https://developers.mercadolibre.com.co/es_ar/autenticacion-y-autorizacion/}{API authentication and authorization guide} for more details. Defaults to System Environment variable "MELI_ACCESS_TOKEN".}
}
\value{
Returns a list of results containing the new test user's user_id, nickname, password, and current status. See the API \href{https://developers.mercadolibre.com.co/es_ar/realiza-pruebas}{documentation} for more information.
}
\description{
Create a test user.
}
