% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{currency_conversion_ratio}
\alias{currency_conversion_ratio}
\title{Get the conversion ratio Mercado Libre uses for two currencies.}
\usage{
currency_conversion_ratio(currency_id1, currency_id2)
}
\arguments{
\item{currency_id1}{Required. currency_id of the first currency of interest.  See mercadolibreR::get_currencies() for valid currency_ids.}

\item{currency_id2}{Required. currency_id of the second currency of interest.  See mercadolibreR::get_currencies() for valid currency_ids.}
}
\value{
Returns the conversion ratio from currency_id1 to currency_id2 as a list().  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get the conversion ratio Mercado Libre uses for two currencies.
}
