% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_category_information}
\alias{get_category_information}
\title{Get category information.}
\usage{
get_category_information(category_id)
}
\arguments{
\item{category_id}{Required. category_id of the category of interest.  See mercadolibreR::get_categories() for valid category ids.}
}
\value{
Returns a list() of the information associated with a given category of products available on Mercado Libre.  Information includes number of items, child categories, and settings.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get category information.
}
