% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_category_promotion_packs}
\alias{get_category_promotion_packs}
\title{Get a category's promotion packs.}
\usage{
get_category_promotion_packs(category_id)
}
\arguments{
\item{category_id}{Required. category_id of the category of interest.  See mercadolibreR::get_categories() for valid category ids.}
}
\value{
Returns a data frame (tibble) of the promotion packs associated with a given category of products.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get a category's promotion packs.
}
