% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_cities}
\alias{get_cities}
\title{Get cities in a specific state in a specific country.}
\usage{
get_cities(country_id, state_id)
}
\arguments{
\item{country_id}{Required. country_id of the country of interest.  See mercadolibreR::get_countries() for valid country_ids.}

\item{stte_id}{Required. state_id of the state of interest.  See mercadolibreR::get_states() for valid state_ids.}
}
\value{
Returns a data frame (tibble) of the cities in a given state.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get cities in a specific state in a specific country.
}
