% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_city_information}
\alias{get_city_information}
\title{Get city information for a specific city.}
\usage{
get_city_information(country_id, state_id, city_id)
}
\arguments{
\item{country_id}{Required. country_id of the country of interest.  See mercadolibreR::get_countries() for valid country_ids.}

\item{state_id}{Required. state_id of the state of interest.  See mercadolibreR::get_states() for valid state_ids.}

\item{city_id}{Required. city_id state of interest.  See mercadolibreR::get_cities() for valid city_ids.}
}
\value{
Returns a list() of information associated with a given city.  Fields may include geo_information and neighborhoods.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get city information for a specific city.
}
