% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_countries}
\alias{get_countries}
\title{Get country information.}
\usage{
get_countries()
}
\value{
Returns a data frame (tibble) of information associated with countries.  Fields include id, name, locale, and currency_id.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get country information.
}
