% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_country_information}
\alias{get_country_information}
\title{Get country information for a specific country.}
\usage{
get_country_information(country_id)
}
\arguments{
\item{country_id}{Required. country_id of the country of interest.  See mercadolibreR::get_countries() for valid country_ids.}
}
\value{
Returns a list() of information associated with a given country.  Fields may include geo_information, states, and time_zone.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get country information for a specific country.
}
