% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_currencies}
\alias{get_currencies}
\title{Get currency information.}
\usage{
get_currencies()
}
\value{
Returns a data frame (tibble) of data regarding Mercado Libre currencies.  Fields include id, symbol, description, and decimal_places.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get currency information.
}
