% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_listing_type_information}
\alias{get_listing_type_information}
\title{Get a listing type's information.}
\usage{
get_listing_type_information(site_id, listing_type_id)
}
\arguments{
\item{site_id}{Required. site_id of the country the product will be listed in.  See mercadolibreR::get_site_ids() for valid site_ids.}

\item{listing_type_id}{Required. listing_type id of the listing of interest.  See mercadolibreR::get_listing_types() for valid listing_type_ids.}
}
\value{
Returns a list() of the information (settings) associated with a listing type.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get a listing type's information.
}
