% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_mercadolibre_access_token}
\alias{get_mercadolibre_access_token}
\title{Get Mercado Libre Access Token}
\usage{
get_mercadolibre_access_token(
  app_id = Sys.getenv("MELI_APP_ID"),
  client_secret = Sys.getenv("MELI_SECRET_KEY")
)
}
\arguments{
\item{app_id}{Defaults to System Environment variable "MELI_APP_ID"}

\item{client_secret}{Defaults to System Environment variable "MELI_SECRET_KEY"}
}
\description{
This function creates a Mercado Libre access token.
}
\examples{
\dontrun{
get_mercadolibre_access_token()
}
}
\keyword{auth}
