% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_site_domains}
\alias{get_site_domains}
\title{Get site domains.}
\usage{
get_site_domains()
}
\value{
Returns a data frame (tibble) of site domains where Mercado Libre is available.  Fields include id, site_id, country_id, locale, and tag.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get site domains.
}
