% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_site_ids}
\alias{get_site_ids}
\title{Get site ids.}
\usage{
get_site_ids()
}
\value{
Returns a data frame (tibble) of site ids where Mercado Libre is available.  Fields include default_currency_id, id, and name.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get site ids.
}
