% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_site_listing_prices}
\alias{get_site_listing_prices}
\title{Get site listing prices.}
\usage{
get_site_listing_prices(site_id, price)
}
\arguments{
\item{site_id}{Required. site_id of the country of interest.  See mercadolibreR::get_site_ids() for valid site_ids.}

\item{price}{Required.  Maximum price for listing and selling.}
}
\value{
Returns a data frame (tibble) of prices for various listing types for the given country's Mercado Libre site.  Fields include listing_type_id, listing_type_name, listing_exposure, requires_picture, currency_id, listing_fee_amount, sale_fee_amount, free_relist, and stop_time.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get site listing prices.
}
