% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{get_site_listing_types}
\alias{get_site_listing_types}
\title{Get site listing types.}
\usage{
get_site_listing_types(site_id)
}
\arguments{
\item{site_id}{Required. site_id of the country of interest.  See mercadolibreR::get_site_ids() for valid site_ids.}
}
\value{
Returns a data frame (tibble) of listing types for the given country's Mercado Libre site.  Fields include site_id, id, and name.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Get site listing types.
}
