% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_states}
\alias{get_states}
\title{Get states in a specific country.}
\usage{
get_states(country_id)
}
\arguments{
\item{country_id}{Required. country_id of the country of interest.  See mercadolibreR::get_countries() for valid country_ids.}
}
\value{
Returns a data frame (tibble) of the states in a given country.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get states in a specific country.
}
