% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_addresses}
\alias{get_user_addresses}
\title{Get user addresses.}
\usage{
get_user_addresses(user_id, authorization = Sys.getenv("MELI_ACCESS_TOKEN"))
}
\arguments{
\item{user_id}{user_id of user.  Use mercadolibreR::get_user_id() if you only have access to the user's nickname and site_id.}

\item{authorization}{A valid access token from developers.mercadolibre.com. See the \href{https://developers.mercadolibre.com.co/es_ar/autenticacion-y-autorizacion/}{API authentication and authorization guide} for more details.}
}
\value{
Returns a list of the user's addresses. See the API \href{https://developers.mercadolibre.com.co/es_ar/usuarios-y-aplicaciones}{documentation} for more information.
}
\description{
Get user addresses.
}
