% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user_id from nickname and site_id.}
\usage{
get_user_id(site_id, nickname)
}
\arguments{
\item{site_id}{Required. site_id of the country in which the user is registered.  See mercadolibreR::get_site_ids() for valid site_ids.}

\item{nickname}{Required. Nickname of the user.}
}
\value{
Returns numeric user_id, or NULL if user is not registered in the given country. See the API \href{https://developers.mercadolibre.com.co/es_ar/producto-consulta-usuarios}{documentation} for more information.
}
\description{
Get user_id from nickname and site_id.
}
