% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_information}
\alias{get_user_information}
\title{Get user information.}
\usage{
get_user_information(
  self = FALSE,
  private_information = TRUE,
  user_id = NULL,
  authorization = NULL
)
}
\arguments{
\item{self}{TRUE if retrieving information about your user; FALSE otherwise.  Defaults to FALSE.}

\item{private_information}{Required if self is FALSE.  TRUE if retrieving private information from another user; FALSE otherwise.  Defaults to FALSE.}

\item{user_id}{Required if self is FALSE.  user_id of user to look up.  Use mercadolibreR::get_user_id() if you only have access to the user's nickname and site_id.}

\item{authorization}{Required if self is TRUE or if private_information is TRUE. If self is TRUE, should be your access token.  If private_information is TRUE, should be the access token for the user from which to retrieve information.  See the \href{https://developers.mercadolibre.com.co/es_ar/autenticacion-y-autorizacion/}{API authentication and authorization guide} for more details.}
}
\value{
Returns a list of information about the user. See the API \href{https://developers.mercadolibre.com.co/es_ar/producto-consulta-usuarios}{documentation} for more information.
}
\description{
Get user information.
}
