% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_zipcode_information}
\alias{get_zipcode_information}
\title{Get zipcode information.}
\usage{
get_zipcode_information(country_id, zipcode)
}
\value{
Returns a list() of information associated with a zipcode.  See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get zipcode information.
}
