% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations_and_currencies.R
\name{get_zipcodes}
\alias{get_zipcodes}
\title{Get all of the zip code numbers between two zip codes for a country.}
\usage{
get_zipcodes(country_id, zipcode_from, zipcode_to)
}
\arguments{
\item{country_id}{Required. country_id of the country of interest.  See mercadolibreR::get_countries() for valid country_ids.}

\item{zipcode_from}{Required. Lower end of range of zipcodes to be returned.}

\item{zipcode_to}{Required. Higher end of range of zipcodes to be returned.}
}
\value{
See the API \href{https://developers.mercadolibre.com.co/es_ar/ubicacion-y-monedas}{documentation} for more information.
}
\description{
Get all of the zip code numbers between two zip codes for a country.
}
