% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercadolibreR.R
\docType{package}
\name{mercadolibreR}
\alias{mercadolibreR}
\title{\code{mercadolibreR} package}
\description{
A Quick and Easy Wrapper for Seller and Listing Data from MercadoLibre's Web API
}
\details{
See the README on
\href{https://github.com/njcorona/mercadolibreR#readme}{GitHub}
}
