% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_and_publications.R
\name{predict_category}
\alias{predict_category}
\title{Predict the category a product should fall under.}
\usage{
predict_category(site_id, title)
}
\arguments{
\item{site_id}{Required. site_id of the country the product will be listed in.  See mercadolibreR::get_site_ids() for valid site_ids.}

\item{title}{Required. Title of the product's listing.}
}
\value{
Returns a list() of the predicted categories with associated ids and probabilities.  See the API \href{https://developers.mercadolibre.com.co/es_ar/categorias-y-publicaciones}{documentation} for more information.
}
\description{
Predict the category a product should fall under.
}
