% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{scopes}
\alias{scopes}
\title{Valid scopes}
\format{
An object of class \code{character} of length 3.
}
\usage{
scopes
}
\description{
Vector of valid scopes for mercadolibreR::get_authorization_code()
}
\keyword{datasets}
