% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_and_search.R
\name{search}
\alias{search}
\title{Search Mercado Libre for item listings.}
\usage{
search(site_id, search, additional_fields = list(), offset = 0, sort = 0)
}
\arguments{
\item{site_id}{Required.  site_id of the country of interest.  See mercadolibreR::get_site_ids() for valid site_ids.}

\item{search}{Required.  What to search for.}

\item{additional_fields}{Optional.  A list() of named character fields that will be added as parameters for the search.  For example, giving list(buying_mode = "buy_it_now") will only search results whose buying_mode is "buy_it_now".}

\item{offset}{Optional.  Default is zero.  Maximum is 1000.  Starting point from which to return records.  For example, to receive the 51st to 100th records, offset = 50.}

\item{sort}{Optional.  Default is zero.  How to sort results.  0 = by relevance; 1 = by price, ascending; 2 = by price, descending.}
}
\value{
Returns a list() of information associated with fifty items listed under the given category.  Fields include sold_quantity, available_quantity, title, permalink, tags, seller, etc.  See the API \href{https://developers.mercadolibre.com.co/es_ar/items-y-busquedas}{documentation} for more information.
}
\description{
Search Mercado Libre for item listings.
}
