% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{set_user_information}
\alias{set_user_information}
\title{Set user information.}
\usage{
set_user_information(user_id, authorization, new_values)
}
\arguments{
\item{user_id}{Required.  user_id of user to update.  Use mercadolibreR::get_user_id() if you only have access to the user's nickname and site_id.}

\item{authorization}{Required.  A valid access token for the user from developers.mercadolibre.com. See the \href{https://developers.mercadolibre.com.co/es_ar/autenticacion-y-autorizacion/}{API authentication and authorization guide} for more details.}

\item{new_values}{a list() of updated information about the user.}
}
\value{
Returns an updated list of information about the user. See the API \href{https://developers.mercadolibre.com.co/es_ar/producto-consulta-usuarios}{documentation} for more information.
}
\description{
Set user information.
}
