% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_and_search.R
\name{sold_quantity}
\alias{sold_quantity}
\title{Returns a tibble() of what sold_quantity values correspond to in reality.  For example, '500' indicates there are 251 to 500 units sold.}
\usage{
sold_quantity()
}
\value{
Returns a tibble() of what sold_quantity values correspond to in reality.  See the API \href{https://developers.mercadolibre.com.co/es_ar/items-y-busquedas}{documentation} for more information.
}
\description{
Returns a tibble() of what sold_quantity values correspond to in reality.  For example, '500' indicates there are 251 to 500 units sold.
}
