% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_result}
\alias{verify_result}
\title{Verify API result}
\usage{
verify_result(res, function_name)
}
\arguments{
\item{res}{API result to check}

\item{function_name}{name of function calling verify_result}
}
\description{
Check API result for error codes
}
